requirements.txt
discord.py
libopus
ffmpeg
	- see Gorialis's comment on https://github.com/Rapptz/discord.py/issues/1149 for help in setting up PATH environement for ffmpeg
youtube-dl
googletrans
